+++ packages/pdfrx/lib/src/wasm/pdfrx_wasm.dart
@@ -112,6 +112,16 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     return await action();
   }
 
+  @override
+  Future<R> compute<M, R>(FutureOr<R> Function(M message) callback, M message) async {
+    throw UnimplementedError('compute() is not implemented for WASM backend.');
+  }
+
+  @override
+  Future<void> stopBackgroundWorker() async {
+    throw UnimplementedError('stopBackgroundWorker() is not implemented for WASM backend.');
+  }
+
   static String? _pdfiumWasmModulesUrlFromMetaTag() {
     final meta = web.document.querySelector('meta[name="pdfium-wasm-module-url"]') as web.HTMLMetaElement?;
     return meta?.content;
@@ -193,6 +203,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     sourceName: sourceName ?? _sourceNameFromData(data),
     passwordProvider: passwordProvider,
     firstAttemptByEmptyPassword: firstAttemptByEmptyPassword,
+    useProgressiveLoading: useProgressiveLoading,
     onDispose: onDispose,
   );
 
@@ -217,6 +228,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     sourceName: 'file%$filePath',
     passwordProvider: passwordProvider,
     firstAttemptByEmptyPassword: firstAttemptByEmptyPassword,
+    useProgressiveLoading: useProgressiveLoading,
     onDispose: null,
   );
 
@@ -259,6 +271,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
         sourceName: 'uri%$uri',
         passwordProvider: passwordProvider,
         firstAttemptByEmptyPassword: firstAttemptByEmptyPassword,
+        useProgressiveLoading: useProgressiveLoading,
         onDispose: cleanupCallbacks,
       );
     } catch (e) {
@@ -272,6 +285,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     required String sourceName,
     required PdfPasswordProvider? passwordProvider,
     required bool firstAttemptByEmptyPassword,
+    required bool useProgressiveLoading,
     required void Function()? onDispose,
   }) async {
     await init();
@@ -298,7 +312,12 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
         throw StateError('Failed to open document: ${result['errorCodeStr']} ($errorCode)');
       }
 
-      return _PdfDocumentWasm._(result, sourceName: sourceName, disposeCallback: onDispose);
+      return _PdfDocumentWasm._(
+        result,
+        sourceName: sourceName,
+        disposeCallback: onDispose,
+        useProgressiveLoading: useProgressiveLoading,
+      );
     }
   }
 
@@ -310,7 +329,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     if (errorCode != null) {
       throw StateError('Failed to create new document: ${result['errorCodeStr']} ($errorCode)');
     }
-    return _PdfDocumentWasm._(result, sourceName: sourceName, disposeCallback: null);
+    return _PdfDocumentWasm._(result, sourceName: sourceName, disposeCallback: null, useProgressiveLoading: false);
   }
 
   @override
@@ -331,7 +350,7 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
     if (errorCode != null) {
       throw StateError('Failed to create document from JPEG data: ${result['errorCodeStr']} ($errorCode)');
     }
-    return _PdfDocumentWasm._(result, sourceName: sourceName, disposeCallback: null);
+    return _PdfDocumentWasm._(result, sourceName: sourceName, disposeCallback: null, useProgressiveLoading: false);
   }
 
   @override
@@ -354,14 +373,25 @@ class PdfrxEntryFunctionsWasmImpl extends PdfrxEntryFunctions {
   }
 
   @override
-  PdfrxBackend get backend => PdfrxBackend.pdfiumWasm;
+  PdfrxBackendType get backendType => PdfrxBackendType.pdfiumWasm;
 }
 
 class _PdfDocumentWasm extends PdfDocument {
-  _PdfDocumentWasm._(this.document, {required super.sourceName, this.disposeCallback})
-    : permissions = parsePermissions(document) {
+  _PdfDocumentWasm._(
+    this.document, {
+    required super.sourceName,
+    required bool useProgressiveLoading,
+    this.disposeCallback,
+  }) : permissions = parsePermissions(document) {
     _pages = parsePages(this, document['pages'] as List<dynamic>);
     updateMissingFonts(document['missingFonts']);
+    if (!useProgressiveLoading) {
+      _notifyDocumentLoadComplete();
+    }
+  }
+
+  void _notifyDocumentLoadComplete() {
+    subject.add(PdfDocumentLoadCompleteEvent(this));
   }
 
   final Map<Object?, dynamic> document;
@@ -446,6 +476,34 @@ class _PdfDocumentWasm extends PdfDocument {
           }
         }
       }
+      // All pages loaded
+      if (firstPageIndex >= pages.length) {
+        _notifyDocumentLoadComplete();
+      }
+    });
+  }
+
+  @override
+  Future<void> reloadPages({List<int>? pageNumbersToReload}) async {
+    if (isDisposed) return;
+    await synchronized(() async {
+      final pageIndices = pageNumbersToReload?.map((n) => n - 1).toList();
+      final result = await _sendCommand(
+        'reloadPages',
+        parameters: {
+          'docHandle': document['docHandle'],
+          if (pageIndices != null) 'pageIndices': pageIndices,
+          'currentPagesCount': pages.length,
+        },
+      );
+      final reloadedPages = parsePages(this, result['pages'] as List<dynamic>);
+      final newPages = pages.toList(growable: false);
+      for (final page in reloadedPages) {
+        newPages[page.pageNumber - 1] = page; // Update the existing page
+      }
+      pages = newPages;
+
+      updateMissingFonts(result['missingFonts']);
     });
   }
 
@@ -599,6 +657,11 @@ class _PdfDocumentWasm extends PdfDocument {
     );
     final bb = result['data'] as ByteBuffer;
     return Uint8List.view(bb.asByteData().buffer, 0, bb.lengthInBytes);
+  }
+
+  @override
+  Future<T> useNativeDocumentHandle<T>(FutureOr<T> Function(int nativeDocumentHandle) task) {
+    throw UnimplementedError('PdfDocument.useNativeDocumentHandle is not implemented for WASM backend.');
   }
 }