+++ packages/pdfrx_engine/lib/src/pdf_document_event.dart
@@ -4,9 +4,14 @@ import 'pdf_page_status_change.dart';
 
 /// PDF document event types.
 enum PdfDocumentEventType {
+  /// [PdfDocumentLoadCompleteEvent]: Document's loading is complete; i.e., all pages are loaded.
+  documentLoadComplete,
+
   /// [PdfDocumentPageStatusChangedEvent]: Page status changed.
   pageStatusChanged,
-  missingFonts, // [PdfDocumentMissingFontsEvent]: Missing fonts changed.
+
+  /// [PdfDocumentMissingFontsEvent]: Missing fonts changed.
+  missingFonts,
 }
 
 /// Base class for PDF document events.
@@ -18,9 +23,20 @@ abstract class PdfDocumentEvent {
   PdfDocument get document;
 }
 
+/// Event that is triggered when the PDF document has finished loading.
+class PdfDocumentLoadCompleteEvent implements PdfDocumentEvent {
+  const PdfDocumentLoadCompleteEvent(this.document);
+
+  @override
+  PdfDocumentEventType get type => PdfDocumentEventType.documentLoadComplete;
+
+  @override
+  final PdfDocument document;
+}
+
 /// Event that is triggered when the status of PDF document pages has changed.
 class PdfDocumentPageStatusChangedEvent implements PdfDocumentEvent {
-  PdfDocumentPageStatusChangedEvent(this.document, {required this.changes});
+  const PdfDocumentPageStatusChangedEvent(this.document, {required this.changes});
 
   @override
   PdfDocumentEventType get type => PdfDocumentEventType.pageStatusChanged;
@@ -31,13 +47,15 @@ class PdfDocumentPageStatusChangedEvent implements PdfDocumentEvent {
   /// The pages that have changed.
   ///
   /// The map is from page number (1-based) to it's status change.
+  ///
+  /// You can assume that the keys in this map are sorted in ascending order.
   final Map<int, PdfPageStatusChange> changes;
 }
 
 /// Event that is triggered when the list of missing fonts in the PDF document has changed.
 class PdfDocumentMissingFontsEvent implements PdfDocumentEvent {
   /// Create a [PdfDocumentMissingFontsEvent].
-  PdfDocumentMissingFontsEvent(this.document, this.missingFonts);
+  const PdfDocumentMissingFontsEvent(this.document, this.missingFonts);
 
   @override
   PdfDocumentEventType get type => PdfDocumentEventType.missingFonts;